% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeforest.R
\name{changeforest}
\alias{changeforest}
\title{Find change points in a time series.}
\usage{
changeforest(X, method, segmentation, control = Control$new())
}
\arguments{
\item{X}{Numerical matrix with time series.}

\item{method}{Either \code{'knn'}, \code{'change_in_mean'} of \code{'random_forest'}
Equal to \code{'random_forest'} by default.}

\item{segmentation}{Either \code{'bs'}, \code{'sbs'} or \code{'wbs'}. Equal to \code{'bs'} 
by default.}

\item{control}{Object of class Control containing hyperparameters. See documentation
of \code{\link{Control}} for details.}
}
\value{
Object of type \code{binary_segmentation_result}.
}
\description{
Find change points in a time series.
}
