% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot.binary_segmentation_result}
\alias{plot.binary_segmentation_result}
\title{Plot an object of class `binary_segmentation_result`.}
\usage{
\method{plot}{binary_segmentation_result}(x, ...)
}
\arguments{
\item{x}{An object of class `binary_segmentation_result`.}

\item{...}{Not used.}
}
\description{
Plot an object of class `binary_segmentation_result`.
}
