% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot.optimizer_result}
\alias{plot.optimizer_result}
\title{Plot an object of class `optimizer_result`.}
\usage{
\method{plot}{optimizer_result}(x, ...)
}
\arguments{
\item{x}{An object of class `optimizer_result`.}

\item{...}{Not used.}
}
\description{
Plot an object of class `optimizer_result`.
}
