% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeforest.R
\name{changeforest}
\alias{changeforest}
\title{Find change points in a time series.}
\usage{
changeforest(X, method, segmentation, control = Control$new())
}
\arguments{
\item{X}{Numerical matrix with time series.}

\item{method}{Either 'knn','change_in_mean' of 'random_forest'.}

\item{segmentation}{Either 'bs', 'sbs' or 'wbs'.}

\item{control}{Object of class Control containing hyperparameters.}
}
\description{
Find change points in a time series.
}
