% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{Control}
\alias{Control}
\title{Storage container for hyperparameters.}
\description{
See rust documentation of changeforest::control::Control for more details.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Control$new()}}
\item \href{#method-clone}{\code{Control$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Control$new(
  minimal_relative_segment_length = NULL,
  minimal_gain_to_split = NULL,
  model_selection_alpha = NULL,
  number_of_wild_segments = NULL,
  seeded_segments_alpha = NULL,
  seed = NULL,
  random_forest_ntrees = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Control$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
